+++ src/core/devices-dag/index.js
@@ -1,0 +1,14 @@
+/**
+ * @file 设备图模块 - 统一导出入口
+ * @module core/devices-dag
+ * @author Zhou Chenyu
+ */
+
+export { DevicesDAG } from "./dag.js";
+export {
+  isPlainObject,
+  isSubDAGDefinition,
+  normalizeHandlerResult,
+} from "./dag-utils.js";
+export { DevicesDAGNode, DevicesDAGEdge } from "./dag-node-edge.js";
+export { DAGBuilder, DAGNodeBuilder, createSubDAG } from "./dag-builder.js";
