+++ src/core/prefixs/tests/drag-anchor-handler.test.js
@@ -1,4 +1,4 @@
-import { DevicesDAG, createSubDAG } from "../../devices/devices-dag.js";
+import { DevicesDAG, createSubDAG } from "../../devices-dag/index.js";
 import { createDragAnchorPrefixHandler } from "../drag-anchor-handler.js";
 
 describe("drag-anchor-handler", () => {
